Moves hello module to be a health module.
Also moves API spec to be in the server directory.
