+++ server/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'java'
     id 'application'
+    id 'org.hidetake.swagger.generator' version '2.14.0'
 }
 
 group 'se.tennander.famlly'
@@ -8,25 +9,37 @@ version '1.0-SNAPSHOT'
 
 sourceCompatibility = 1.8
 
-mainClassName = "se.tennander.family.Main"
+mainClassName = "se.tennander.family.FamilyServer"
 
 repositories {
     mavenCentral()
+    jcenter()
 }
 
 dependencies {
+    swaggerUI 'org.webjars:swagger-ui:3.10.0'
+
     compile group: 'com.google.inject', name: 'guice', version: '4.0'
     compile group: 'com.google.inject.extensions', name: 'guice-multibindings', version: '4.2.2'
     compile "com.fasterxml.jackson.core:jackson-databind:2.9.6"
     compile group: 'io.javalin', name: 'javalin', version: '2.4.0'
+
     testCompile group: 'junit', name: 'junit', version: '4.12'
+
+    runtime "org.slf4j:slf4j-simple:1.7.25"
 }
 
-task swggerUi(type: Copy) {
-    from tasks.getByPath(":api:generateSwaggerUIFamily").outputs into "build/res/static/docs"
+swaggerSources {
+    family {
+        inputFile = file('api.yml')
+    }
+}
+
+task swaggerUi(type: Copy) {
+    from tasks.getByPath("generateSwaggerUIFamily").outputs into "build/res/static/docs"
 }
 
 processResources {
-    dependsOn("swggerUi")
+    dependsOn("swaggerUi")
     from "build/res/"
 }
