+++ server/src/main/java/se/tennander/family/Service.java
@@ -2,22 +2,29 @@ package se.tennander.family;
 
 import io.javalin.apibuilder.EndpointGroup;
 
+/**
+ * A service public facing API.
+ * <p>
+ * It is one of the route URIs accessible from the api.
+ */
 public interface Service {
 
   /**
-   * Returns the service name
-   * @return
+   * Returns the service name.
+   *<p>
+   * The name will be prefixed on the URI to the service.
+   * @return The name of the service.
    */
-  public String serviceName();
+  String serviceName();
 
   /**
    * Wire the service routes.
-   * @param route
+   * @param wiring Wiring object allowing services to wire end points.
    */
-  public void wire(Routes route);
+  void wire(Wiring wiring);
 
   @FunctionalInterface
-  public interface Routes {
+  interface Wiring {
     void addRoute(String route, EndpointGroup endpointGroup);
   }
 }
