+++ server/src/main/java/se/tennander/family/healthchecks/HealthCheckModule.java
@@ -1,13 +1,13 @@
-package se.tennander.family.hello;
+package se.tennander.family.healthchecks;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.multibindings.Multibinder;
 import se.tennander.family.Service;
 
-public class HelloModule extends AbstractModule {
+public class HealthCheckModule extends AbstractModule {
   @Override
   protected void configure() {
     Multibinder<Service> servicesBinder = Multibinder.newSetBinder(binder(), Service.class);
-    servicesBinder.addBinding().to(World.class);
+    servicesBinder.addBinding().to(HealthChecks.class);
   }
 }
