+++ settings.gradle
@@ -2,5 +2,4 @@ rootProject.name = 'family'
 
 include 'server'
 include 'client'
-include 'api'
